- FIXED: The empty list literal wasn't being intialized, resulting in the empty list "()" causing a crash. - FIXED: Design bug in block instancing system that prevented recursion from working correctly. - FIXED: Fixed all MSVC++ warnings on warning level 3.

+++ trunk/include/Slib-List.hpp
@@ -72,7 +72,7 @@ public:
 	VariableBasePointer Operate( VariableBasePointer );
 	
 private:
-	ListPointer QuickSort( ListPointer, unsigned int Begin, unsigned int End );
+	ListPointer QuickSort( ListPointer, unsigned long Begin, unsigned long End );
 	
 	virtual int Compare( VariableBasePointer x, VariableBasePointer y );
